Exposed `preserveFocus` on OutputChannel#show.
From now on, calling `show` with `preserveFocus` `true` will ensure that
the Output widget is revealed but not activated.
The default behavior remains `preserveFocus` `false` which will reveal
and activate the widget.

Removed `OutputWidget#setInput`; it is automatically wired into `show`.
